+++ settings.py.bak
@@ -162,6 +162,7 @@ INSTALLED_APPS = [
     'paciente',
     'obra_social',
     'medico',
+    'anestesista',
     'turno',
     'estudio',
     'comprobante',
